% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_interval_data.R
\name{fit_interval_data}
\alias{fit_interval_data}
\title{Fitting Data for Interval Calculation}
\usage{
fit_interval_data(
  interval_data,
  distribution = c("gamma", "gamma-trunc"),
  stan_opts = list(),
  return_raw = FALSE
)
}
\arguments{
\item{interval_data}{a named list containing the interval data}

\item{distribution}{a string vector indicating the distributions with which
to fit the data}

\item{stan_opts}{a named list with additional options passed to CmdStanR}

\item{return_raw}{a logical, should  the samples be returned}
}
\description{
This function takes data from the \code{prep_interval_data} function
and fits it to a series of speficied distributions
}
