% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_interval_data.R
\name{prep_interval_data}
\alias{prep_interval_data}
\title{Prepare Interval Data}
\usage{
prep_interval_data(dat)
}
\arguments{
\item{dat}{a data.frame with columes named er, el sl,sr}
}
\value{
a named list for use with \code{fit_interval_data}
}
\description{
This function can be used to prepare data for fitting
}
